# src/v2/docs/openapi.yaml
openapi: 3.0.0
info:
  title: TMJApp API V2
  version: 2.0.0
  description: API v2 organizada por visoes (admin, passenger, driver).
servers:
  - url: http://localhost:3000/api/v2
  - url: http://api.tmjapp.com.br/api/v2
...
paths:
...
  /auth/login:
    post:
      tags: [Auth V2]
      summary: Login do usuario (email/telefone/identifier)
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
            examples:
              porEmail:
                summary: Login com email no campo identifier
                value:
                  identifier: "usuario@exemplo.com"
                  password: "Senha@123"
              porTelefone:
                summary: Login com telefone no campo identifier
                value:
                  identifier: "81999999999"
                  password: "Senha@123"
      responses:
        '200':
          description: Login realizado com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginSuccessResponse'
        '400':
          description: Dados obrigatorios ausentes
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        '401':
          description: Credenciais invalidas
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
...
  /auth/onboarding-status/{id}:
    get:
      tags: [Auth V2]
      summary: Consultar onboarding por id (admin ou proprio usuario)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Status retornado com sucesso
          content:
            application/json:
              schema:
                type: object
                properties:
                  onboardingStatus:
                    $ref: '#/components/schemas/OnboardingStatus'
        '401':
          description: Token ausente/invalido
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        '403':
          description: Acesso negado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
  /auth/email/send-code:
    post:
      tags: [Auth V2]
      summary: Enviar codigo de validacao para o email do usuario autenticado
      responses:
        '200':
          description: Codigo enviado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SendCodeResponse'
        '400':
          description: Falha ao enviar codigo
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
  /auth/email/verify:
    post:
      tags: [Auth V2]
      summary: Validar email com codigo
      description: Validacao de email ativa o usuario e marca o telefone como validado temporariamente enquanto a verificacao por SMS estiver desabilitada.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VerifyCodeRequest'
      responses:
        '200':
          description: Email validado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationSuccessResponse'
        '400':
          description: Codigo invalido
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
...
  /auth/phone/send-code:
    post:
      tags: [Auth V2]
      summary: Enviar codigo de validacao para o celular do usuario autenticado via AWS SNS
      description: Endpoint mantido por compatibilidade. A ativacao do usuario nao depende temporariamente de SMS.
      responses:
        '200':
          description: Codigo enviado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SendCodeResponse'
        '400':
          description: Falha ao enviar codigo
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
  /auth/phone/verify:
    post:
      tags: [Auth V2]
      summary: Validar celular com codigo
      description: Endpoint mantido por compatibilidade. A ativacao do usuario nao depende temporariamente de SMS.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VerifyCodeRequest'
      responses:
        '200':
          description: Celular validado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationSuccessResponse'
        '400':
          description: Codigo invalido
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
...
  /auth/forgot-password:
    post:
      tags: [Auth V2]
      summary: Recuperar senha
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: Operação realizada com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessMessageResponse'
        '400':
          description: Erro de validação ou regra de negócio
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
  /auth/forgot-password/verify:
    post:
      tags: [Auth V2]
      summary: Verificar código de recuperação de senha
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: Operação realizada com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessMessageResponse'
        '400':
          description: Erro de validação ou regra de negócio
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
  /auth/reset-password:
    post:
      tags: [Auth V2]
      summary: Redefinir senha
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: Operação realizada com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessMessageResponse'
        '400':
          description: Erro de validação ou regra de negócio
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
  /auth/resend-code:
    post:
      tags: [Auth V2]
      summary: Reenviar código
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: Operação realizada com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessMessageResponse'
        '400':
          description: Erro de validação ou regra de negócio
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  schemas:
...
    VerifyCodeRequest:
      type: object
      required: [code]
      properties:
        code:
          type: string
    SendCodeResponse:
      type: object
      required: [success, message, channel, sentTo]
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
        channel:
          type: string
          enum: [EMAIL, PHONE]
        sentTo:
          type: string
        sentAt:
          type: string
          format: date-time
          nullable: true
...
    ValidationSuccessResponse:
      type: object
      required: [success, message, token, user]
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
        token:
          type: string
          description: JWT atualizado apos validacao para refletir o novo authStatus.
        user:
          $ref: '#/components/schemas/AuthenticatedUser'
        onboardingStatus:
          nullable: true
          oneOf:
            - $ref: '#/components/schemas/OnboardingStatus'
            - type: 'null'
    LoginRequest:
      type: object
      required: [password]
      properties:
        identifier:
          type: string
          description: Campo recomendado para email ou telefone.
        email:
          type: string
          format: email
          description: Campo alternativo legado.
        phone:
          type: string
          description: Campo alternativo legado.
        password:
          type: string
      anyOf:
        - required: [identifier]
        - required: [email]
        - required: [phone]
    LoginSuccessResponse:
      type: object
      required: [success, token, user]
      properties:
        success:
          type: boolean
          example: true
        token:
          type: string
          description: JWT do usuario autenticado.
        user:
          $ref: '#/components/schemas/AuthenticatedUser'
        onboardingStatus:
          nullable: true
          description: Presente para perfil motorista/admin.
          oneOf:
            - $ref: '#/components/schemas/OnboardingStatus'
            - type: 'null'
    OnboardingStatus:
      type: object
      required: [isCompleted, isUnderReview, canDrive, steps]
      properties:
        isCompleted:
          type: boolean
        isUnderReview:
          type: boolean
        canDrive:
          type: boolean
        steps:
          type: object
          properties:
            profilePhoto:
              $ref: '#/components/schemas/OnboardingStepWithReview'
            email:
              $ref: '#/components/schemas/OnboardingStep'
            phone:
              $ref: '#/components/schemas/OnboardingStep'
            documents:
              $ref: '#/components/schemas/OnboardingStepWithReview'
            criminalRecord:
              $ref: '#/components/schemas/OnboardingStepWithReview'
            vehicle:
              $ref: '#/components/schemas/OnboardingStepWithReview'
            vehiclePhotos:
              $ref: '#/components/schemas/OnboardingStepWithReview'
    OnboardingStep:
      type: object
      properties:
        completed:
          type: boolean
    OnboardingStepWithReview:
      type: object
      properties:
        completed:
          type: boolean
        underReview:
          type: boolean
...
    SuccessMessageResponse:
      type: object
      required: [success, message]
      properties:
        success:
          type: boolean
        message:
          type: string
...
    AuthenticatedUser:
      type: object
      required:
        [id, name, email, phone, role, authStatus, emailVerified, phoneVerified]
      properties:
        id:
          type: string
        name:
          type: string
        email:
          type: string
          format: email
        phone:
          type: string
        cpfMasked:
          type: string
          nullable: true
          description: CPF mascarado para exibição segura na UI.
        role:
          type: string
          enum: [passenger, driver, admin]
        authStatus:
          type: string
          enum: [PENDING_EMAIL, PENDING_PHONE, ACTIVE, BLOCKED]
          description: "PENDING_PHONE permanece apenas por compatibilidade legada. O fluxo atual promove para ACTIVE assim que o email e validado."
        emailVerified:
          type: boolean
        phoneVerified:
          type: boolean
    ApiErrorResponse:
      type: object
      required: [success, message]
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
        code:
          type: string
        details:
          type: object
          nullable: true
          additionalProperties: true
        traceId:
          type: string
security:
  - bearerAuth: []